issue 16

+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/ui/navigation/NavigationComponent.kt
@@ -62,7 +62,7 @@ fun NavigationComponent(
             val publicationId = navBackStackEntry.arguments?.getString("publicationId") ?: ""
             PublicationDetailsScreen(
                 controller = navigationController,
-                publicationId = publicationId.toLong(),
+                publicationId = publicationId,
             )
         }
     }
